Move float-related API to the correct header file.
